pyproject.toml
[tool.poetry]
name = "airbnb"
version = "0.1.0"
description = ""
authors = ["Youngjin Jo <yjinjo@berkeley.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.164.0"}
strawberry-graphql-django = "^0.9.2"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
